ActorVars.xml restructure; Began simplifying the SceneHeader parser

+++ ZActor/MActors/MSimpleActor.cs
@@ -2,7 +2,7 @@
 {
     class MSimpleActor : MActorRecord
     {
-        public MSimpleActor(byte[] record)
+        public MSimpleActor(short[] record)
             : base(record)
         {
             //"??? (unlisted)"
