+++ ZActor/OActors/BombflowerActor.cs
@@ -2,7 +2,7 @@
 {
     class BombflowerActor:ActorRecord_Wrapper
     {
-        public BombflowerActor(byte[] record, params int[] p)
+        public BombflowerActor(short[] record, params int[] p)
             : base(record)
         {
             objectDependencies = p;
