+++ ZActor/OActors/RisingStonePlatformActor.cs
@@ -2,7 +2,7 @@
 {
     class RisingStonePlatformActor : ActorRecord_Wrapper
     {
-        public RisingStonePlatformActor(byte[] record, params int[] p)
+        public RisingStonePlatformActor(short[] record, params int[] p)
             : base(record)
         {
             objectDependencies = p;
