+++ source/SceneRoom/Actor/LinkActor.cs
@@ -7,7 +7,7 @@ namespace mzxrules.OcaLib.Actor
 {
     public class LinkActor : ActorRecord
     {
-        public LinkActor(byte[] data) : base(data) { }
+        public LinkActor(short[] data) : base(data) { }
         public LinkActor() { }
         
         protected override string GetActorName()
